swapped to pycord, started on speech translation
speech translation / transcription does not work at the moment. working theory is it is an issue with how to .wav file is being saved as one generated by audacity works but audacity cannot read the one made by the bot. somehow VS code is able to read the possibly corrupted wav file though.